modules/HistoricalDataCollector/Ctrader/requirements.txt
ctrader-open-api==0.9.2
requests
Twisted
pyOpenSSL
redis
psycopg2-binary